refactor: move generic functions to top-level
