+++ pydarts/gui/app.py
@@ -1,9 +1,9 @@
 import tkinter as tk
-from typing import Iterable
 
 import customtkinter as ctk
 
 import pydarts
+import pydarts.gui
 from pydarts.gui import BaseStage
 from pydarts.gui.pregame import PregameStage
 
@@ -36,7 +36,7 @@ class App(ctk.CTk):
         return None
 
     def _enable_debug(self):
-        highlight_children(self)
+        pydarts.gui.highlight_children(self)
         sequences = []
         for sequence in sequences:
             self.bind_all(sequence, self._handle_event)
@@ -47,23 +47,3 @@ class App(ctk.CTk):
             f"  event: {event!r}\n"
             f"  widget: {event.widget!r}"
         )
-
-
-def highlight_children(parent: tk.Misc):
-    for widget in walk_children(parent):
-        try:
-            widget.configure(borderwidth=1, relief="solid")  # type: ignore
-        except (tk.TclError, ValueError):
-            # widget does not support borders, fine
-            pass
-
-
-def walk_children(parent: tk.Misc, max_depth: int = -1) -> Iterable[tk.Misc]:
-    children = [(child, max_depth) for child in reversed(parent.winfo_children())]
-    while children:
-        child, max_depth = children.pop()
-        yield child
-        if max_depth == 0:
-            continue
-        for grandchild in reversed(child.winfo_children()):
-            children.append((grandchild, max_depth - 1))
